requirements.txt
# Core
fastapi==0.135.1
uvicorn==0.42.0
starlette==0.52.1
pydantic==2.12.5
pydantic_core==2.41.5

# Database
psycopg2-binary>=2.9.10

# AI
groq==1.1.1

# HTTP
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9

# Auth
PyJWT>=2.8.0

# Rate limiting
slowapi>=0.1.9

# Data processing
pandas==2.2.3
pyarrow==19.0.1

# Environment
python-dotenv==1.0.1

# Server (production)
gunicorn>=21.2.0

# Utils
tqdm==4.67.3
colorama==0.4.6
certifi==2026.2.25
idna==3.11
urllib3==2.6.3
sniffio==1.3.1
h11==0.16.0
click==8.3.1
typing_extensions==4.15.0
typing-inspection==0.4.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
charset-normalizer==3.4.6
distro==1.9.0
cffi==2.0.0
cryptography==46.0.5
pycparser==3.0
